Moved baseUrl into src
